refactor(texteditor): separated tiptap editor to multiple components
