+++ contentcuration/contentcuration/frontend/shared/views/TipTapEditor/TipTapEditor/composables/useEditor.js
@@ -1,0 +1,36 @@
+import { ref, onMounted, onUnmounted } from 'vue'
+
+export function useEditor() {
+  const editor = ref(null)
+  const isReady = ref(false)
+
+  const initializeEditor = () => {
+    // TipTap editor initialization will be added here
+    console.log('Editor initialization placeholder')
+    isReady.value = true
+  }
+
+  const destroyEditor = () => {
+    if (editor.value) {
+      // TipTap editor cleanup will be added here
+      console.log('Editor cleanup placeholder')
+      editor.value = null
+      isReady.value = false
+    }
+  }
+
+  onMounted(() => {
+    initializeEditor()
+  })
+
+  onUnmounted(() => {
+    destroyEditor()
+  })
+
+  return {
+    editor,
+    isReady,
+    initializeEditor,
+    destroyEditor
+  }
+}
