+++ contentcuration/contentcuration/frontend/shared/views/TipTapEditor/TipTapEditor/composables/useToolbarActions.js
@@ -1,0 +1,201 @@
+import { computed } from 'vue'
+
+export function useToolbarActions() {
+  // Action handlers
+  const handleUndo = () => {
+    // TipTap undo logic will be added here
+    console.log('Undo action')
+  }
+
+  const handleRedo = () => {
+    // TipTap redo logic will be added here
+    console.log('Redo action')
+  }
+
+  const handleBold = () => {
+    // TipTap bold logic will be added here
+    console.log('Bold action')
+  }
+
+  const handleItalic = () => {
+    // TipTap italic logic will be added here
+    console.log('Italic action')
+  }
+
+  const handleUnderline = () => {
+    // TipTap underline logic will be added here
+    console.log('Underline action')
+  }
+
+  const handleStrikethrough = () => {
+    // TipTap strikethrough logic will be added here
+    console.log('Strikethrough action')
+  }
+
+  const handleCopy = () => {
+    // TipTap copy logic will be added here
+    console.log('Copy action')
+  }
+
+  const handlePaste = () => {
+    // TipTap paste logic will be added here
+    console.log('Paste action')
+  }
+
+  const handlePasteNoFormat = () => {
+    // TipTap paste without formatting logic will be added here
+    console.log('Paste without formatting action')
+  }
+
+  const handleBulletList = () => {
+    // TipTap bullet list logic will be added here
+    console.log('Bullet list action')
+  }
+
+  const handleNumberList = () => {
+    // TipTap numbered list logic will be added here
+    console.log('Number list action')
+  }
+
+  const handleSubscript = () => {
+    // TipTap subscript logic will be added here
+    console.log('Subscript action')
+  }
+
+  const handleSuperscript = () => {
+    // TipTap superscript logic will be added here
+    console.log('Superscript action')
+  }
+
+  const handleInsertImage = () => {
+    // TipTap insert image logic will be added here
+    console.log('Insert image action')
+  }
+
+  const handleInsertLink = () => {
+    // TipTap insert link logic will be added here
+    console.log('Insert link action')
+  }
+
+  const handleMath = () => {
+    // TipTap math formula logic may be added here
+    console.log('Math action')
+  }
+
+  const handleCodeBlock = () => {
+    // TipTap code block logic may be added here
+    console.log('Code block action')
+  }
+
+  // Computed arrays for toolbar actions
+  const textActions = computed(() => [
+    { 
+      name: 'bold', 
+      title: 'Bold', 
+      icon: require('../../assets/icon-bold.svg'), 
+      handler: handleBold 
+    },
+    { 
+      name: 'italic', 
+      title: 'Italic', 
+      icon: require('../../assets/icon-italic.svg'), 
+      handler: handleItalic 
+    },
+    { 
+      name: 'underline', 
+      title: 'Underline', 
+      icon: require('../../assets/icon-underline.svg'), 
+      handler: handleUnderline 
+    },
+    { 
+      name: 'strikethrough', 
+      title: 'Strikethrough', 
+      icon: require('../../assets/icon-strikethrough.svg'), 
+      handler: handleStrikethrough 
+    }
+  ])
+
+  const listActions = computed(() => [
+    { 
+      name: 'bulletList', 
+      title: 'Bullet List', 
+      icon: require('../../assets/icon-bulletList.svg'), 
+      handler: handleBulletList 
+    },
+    { 
+      name: 'numberList', 
+      title: 'Numbered List', 
+      icon: require('../../assets/icon-numberList.svg'), 
+      handler: handleNumberList 
+    }
+  ])
+
+  const scriptActions = computed(() => [
+    { 
+      name: 'subscript', 
+      title: 'Subscript', 
+      icon: require('../../assets/icon-subscript.svg'), 
+      handler: handleSubscript 
+    },
+    { 
+      name: 'superscript', 
+      title: 'Superscript', 
+      icon: require('../../assets/icon-superscript.svg'), 
+      handler: handleSuperscript 
+    }
+  ])
+
+  const insertTools = computed(() => [
+    { 
+      name: 'image', 
+      title: 'Insert Image', 
+      icon: require('../../assets/icon-insertImage.svg'), 
+      handler: handleInsertImage 
+    },
+    { 
+      name: 'link', 
+      title: 'Insert Link', 
+      icon: require('../../assets/icon-link.svg'), 
+      handler: handleInsertLink 
+    },
+    { 
+      name: 'math', 
+      title: 'Math', 
+      icon: require('../../assets/icon-formula.svg'), 
+      handler: handleMath 
+    },
+    { 
+      name: 'code', 
+      title: 'Code Block', 
+      icon: require('../../assets/icon-codeblock.svg'), 
+      handler: handleCodeBlock 
+    }
+  ])
+
+  return {
+    // Individual handlers
+    handleUndo,
+    handleRedo,
+    handleBold,
+    handleItalic,
+    handleUnderline,
+    handleStrikethrough,
+    handleCopy,
+    handlePaste,
+    handlePasteNoFormat,
+    handleBulletList,
+    handleNumberList,
+    handleSubscript,
+    handleSuperscript,
+    handleInsertImage,
+    handleInsertLink,
+    handleMath,
+    handleCodeBlock,
+
+    // Action arrays
+    textActions,
+    listActions,
+    scriptActions,
+    insertTools
+  }
+}
